:recycle: Move `test_patch_predictor_new` to `test_patch_predictor`
